requirements.txt
boto3>=1.20.21
sagemaker~=2.131.0
    # via -r requirements.in
sagemaker[local]~=2.131.0
smdebug-rulesconfig
pandas~=1.5.1
sklearn~=0.0
seaborn
matplotlib
numpy~=1.23.5
six~=1.15
testresources
xgboost~=1.4.2
scikit-learn~=0.24.2
awswrangler~=2.16.1
botocore~=1.24.44
requests~=2.26.0
pytest~=7.1.1
setuptools~=45.2.0
snowflake-connector-python==3.0.3
snowflake-snowpark-python==1.4.0
snowflake-sqlalchemy==1.4.7
sqlalchemy==1.4.47